Added small features at compile.sh
Added usage menu ```./compile.sh --help```

+++ plugins/compile.sh
@@ -7,17 +7,51 @@
 #
 # This file is part of AMX Mod X.
 
-# to compile all source files in the directory
-#  execute the script without any parameters
-# to compile specific or multiple source files
-#  execute the script with parameters set as source files' name
-#  e.g ./compile.sh admin.sma admincmd.sma adminchat.sma
-
 test -e compiled || mkdir compiled
 rm -f temp.txt
 
 if [ $# != 0 ]
 	then
+		case $1 in
+			"--help" | "-h" )
+				echo "  Usage:"
+				echo "--help    (-h) : This menu"
+				echo "--version (-v) : AMX MOD X Compiler version"
+				echo -e "--latest  (-l) : Latest AMX MOD X version\n"
+				echo "To compile all source files in the directory"
+				echo "  execute the script without any parameters"
+				echo "To compile specific or multiple source files"
+				echo "  execute the script with parameters set as source files' name"
+				echo "  e.g ./compile.sh admin.sma admincmd.sma adminchat.sma"
+				;;
+			 "--version" | "-v" )
+				./amxxpc -q | head -n 3
+	 			;;
+	 		"--latest" | "-l" )
+				echo "Latest AMX MOD X version"
+				os="$(uname -s)"
+				
+				case "$os" in
+					"Linux"  ) os="linux" ;;
+					"Darwin" ) os="mac" ;;
+				*)
+				esac
+				
+				latest=$(wget -qO- https://www.amxmodx.org/amxxdrop/1.9/amxmodx-latest-base-$os)
+				
+				if [ $? -eq 0 ]
+				then
+					echo -n "1.9 Build: "
+					echo $latest | grep -o -P '(?<=git).*(?=-base)'
+					echo -n "1.10 Build: "
+					wget -qO- https://www.amxmodx.org/amxxdrop/1.10/amxmodx-latest-base-$os | grep -o -P '(?<=git).*(?=-base)'
+				else
+					echo "Internet connection is required"
+				fi
+				;;
+			*)
+		esac
+		
 		for src in "$@"
 		do
 			if [[ $src == *.sma ]]
@@ -38,5 +72,8 @@ if [ $# != 0 ]
 		done
 fi
 
-less temp.txt
-rm temp.txt
+if [ -f temp.txt ]
+then
+	less temp.txt
+	rm temp.txt
+fi
